Move disabled IRQ guards to a submodule
